.mcp/requirements.txt
fastapi>=0.104.1
fastapi-mcp>=0.1.0
pydantic>=2.5.0
uvicorn[standard]>=0.24.0
supabase>=2.3.0
mindsdb-sdk>=1.6.0
python-dotenv>=1.0.0
